requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
blinker==1.6.2
    # via flask
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via flask
flask==2.3.2
    # via -r requirements.in
grpcio==1.54.2
    # via grpcio-tools
grpcio-tools==1.54.2
    # via -r requirements.in
idna==3.4
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
pillow==9.5.0
    # via reportlab
protobuf==4.23.2
    # via grpcio-tools
pydantic==1.10.9
    # via -r requirements.in
python-status==1.0.1
    # via -r requirements.in
reportlab==4.0.4
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
typing-extensions==4.6.3
    # via pydantic
urllib3==2.0.3
    # via requests
werkzeug==2.3.6
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
